requirements.txt
albumentations==1.3.0
einops==0.5.0
elasticdeform==0.5.0
eyepy==0.0.0a0
h5py==3.7.0
matplotlib==3.6.2
numpy==1.23.4
patchify==0.2.3
Pillow==9.3.0
PySide2==5.15.2.1
pytorch_model_summary==0.1.2
PyYAML==6.0
scikit_learn==1.1.3
scipy==1.9.3
skimage==0.0
timm==0.6.11
torch==1.10.1
torchsummary==1.5.1
torchvision==0.14.0
tqdm==4.64.1
